+++ classes/add_NMEA_0183.py
@@ -15,15 +15,16 @@
 # You should have received a copy of the GNU General Public License
 # along with Openplotter. If not, see <http://www.gnu.org/licenses/>.
 import wx, pynmea2, inspect, webbrowser, json, subprocess, requests, re
-from classes.paths import Paths
 
 class addNMEA_0183(wx.Dialog):
 
-	def __init__(self,edit):
+	def __init__(self,edit,conf):
+
+		self.conf = conf
+		self.home = self.conf.home
+		self.currentpath = self.home+self.conf.get('GENERAL', 'op_folder')+'/openplotter'
 
 		wx.Dialog.__init__(self, None, title=_('add NMEA 0183 sentence'), size=(670,410))
-
-		self.paths=Paths()
 
 		panel = wx.Panel(self)
 
@@ -109,7 +110,9 @@ class addNMEA_0183(wx.Dialog):
 
 		else:
 			self.button_rmc_examp.Disable()
-			self.rate.SetValue(str(edit[1][2]))
+			for rateI in self.rate_list:
+				if float(rateI) == edit[1][2]:		
+					self.rate.SetValue(rateI)
 			self.sentence.SetValue('$--'+edit[1][0])
 			self.nmea=edit[1]
 			for index,item in enumerate(self.sentences):
@@ -134,57 +137,57 @@ class addNMEA_0183(wx.Dialog):
 		
 	def onEdit_fields(self,e):
 		selected_field=self.list_fields.GetFirstSelected()
-		#print self.list_fiel[selected_field]
-		if type(self.list_fiel[selected_field][2]) is list:
-			if (self.list_fiel[selected_field][2][2] == '+') == True and self.list_fiel[selected_field][2][3] == 0:
-				self.check_value_type_(self.list_value_type[2])
-				self.skvessels.SetValue(self.list_vessels[1])
-				self.onSelect_vessel_(self.list_vessels[1])
-				sk = self.list_fiel[selected_field][2][0]
-				sk0 = sk.split('.')[0]
-				skr = sk[len(sk0)+1:]
-				self.skgroups.SetValue(sk0)
-				self.onSelect_group_(sk0)
-				self.signalk.SetValue(skr)
-				format = self.list_formats[0]
-				formats = self.list_fiel[selected_field][2][1]
-				for value_ in self.list_formats:
-					if formats in value_:
-						format = value_
-						break						
-				self.formats.SetValue(format)
+		if len(self.list_fiel) != 0:
+			if type(self.list_fiel[selected_field][2]) is list:
+				if (self.list_fiel[selected_field][2][2] == '+') == True and self.list_fiel[selected_field][2][3] == 0:
+					self.check_value_type_(self.list_value_type[2])
+					self.skvessels.SetValue(self.list_vessels[1])
+					self.onSelect_vessel_(self.list_vessels[1])
+					sk = self.list_fiel[selected_field][2][0]
+					sk0 = sk.split('.')[0]
+					skr = sk[len(sk0)+1:]
+					self.skgroups.SetValue(sk0)
+					self.onSelect_group_(sk0)
+					self.signalk.SetValue(skr)
+					format = self.list_formats[0]
+					formats = self.list_fiel[selected_field][2][1]
+					for value_ in self.list_formats:
+						if formats in value_:
+							format = value_
+							break						
+					self.formats.SetValue(format)
+				else:
+					self.check_value_type_(self.list_value_type[1])
+					self.skvessels.SetValue(self.list_vessels[1])
+					self.onSelect_vessel_(self.list_vessels[1])
+					sk = self.list_fiel[selected_field][2][0]
+					sk0 = sk.split('.')[0]
+					skr = sk[len(sk0)+1:]
+					self.skgroups.SetValue(sk0)
+					self.onSelect_group_(sk0)
+					self.signalk.SetValue(skr)
+					opera = self.list_operators[0]
+					for value_ in self.list_operators:
+						if self.list_fiel[selected_field][2][2] in value_:
+							opera = value_
+							break						
+					self.operator.SetValue(opera)
+					format = self.list_formats[0]
+					formats = self.list_fiel[selected_field][2][1]
+					for value_ in self.list_formats:
+						if formats in value_:
+							format = value_
+							break						
+					self.formats.SetValue(format)
+					self.string_number.SetValue(str(self.list_fiel[selected_field][2][3]))
 			else:
-				self.check_value_type_(self.list_value_type[1])
-				self.skvessels.SetValue(self.list_vessels[1])
-				self.onSelect_vessel_(self.list_vessels[1])
-				sk = self.list_fiel[selected_field][2][0]
-				sk0 = sk.split('.')[0]
-				skr = sk[len(sk0)+1:]
-				self.skgroups.SetValue(sk0)
-				self.onSelect_group_(sk0)
-				self.signalk.SetValue(skr)
-				opera = self.list_operators[0]
-				for value_ in self.list_operators:
-					if self.list_fiel[selected_field][2][2] in value_:
-						opera = value_
-						break						
-				self.operator.SetValue(opera)
-				format = self.list_formats[0]
-				formats = self.list_fiel[selected_field][2][1]
-				for value_ in self.list_formats:
-					if formats in value_:
-						format = value_
-						break						
-				self.formats.SetValue(format)
-				self.string_number.SetValue(str(self.list_fiel[selected_field][2][3]))
-		else:
-			if self.is_number(self.list_fiel[selected_field][2]):
-				self.check_value_type_(self.list_value_type[3])
-				self.string_number.SetValue(self.list_fiel[selected_field][2])
-			else:
-				if len(self.list_fiel[selected_field][2])>0:
-					self.check_value_type_(self.list_value_type[4])
+				if self.is_number(self.list_fiel[selected_field][2]):
+					self.check_value_type_(self.list_value_type[3])
 					self.string_number.SetValue(self.list_fiel[selected_field][2])
+				else:
+					if len(self.list_fiel[selected_field][2])>0:
+						self.check_value_type_(self.list_value_type[4])
+						self.string_number.SetValue(self.list_fiel[selected_field][2])
 			
 
 	def onSelect_vessel(self,e):
@@ -277,7 +280,7 @@ class addNMEA_0183(wx.Dialog):
 		operator=self.operator.GetValue()
 		str_num=self.string_number.GetValue()
 		if selected_type == _('String'):
-			if not re.match('^[0-9a-zA-Z/]+$', str_num) and len(str_num) > 0 :
+			if not re.match('^[0-9a-zA-Z/_]+$', str_num) and len(str_num) > 0 :
 				self.list_fields.SetStringItem(selected_field,1,_('Failed. String must contain only allowed characters.'))
 				str_num=''
 				self.string_number.SetValue('')
@@ -350,7 +353,7 @@ class addNMEA_0183(wx.Dialog):
 		self.nmea[1][selected_field]=''
 
 	def nmea_info(self, e):
-		url = self.paths.op_path+'/docs/NMEA.html'
+		url = self.currentpath+'/docs/NMEA.html'
 		webbrowser.open(url,new=2)
 		
 	def rmc_examp(self, e):
@@ -373,7 +376,7 @@ class addNMEA_0183(wx.Dialog):
 	def check_value_type_(self, selected):
 		self.value_type.SetValue(selected)
 		#none
-		if not selected or selected==self.list_value_type[0]:
+		if not selected or selected.encode('utf-8')==self.list_value_type[0]:
 			self.skvessels.Disable()
 			self.skgroups.Disable()
 			self.signalk.Disable()