+++ tools/NMEA_0183_generator.py
@@ -15,27 +15,29 @@
 # You should have received a copy of the GNU General Public License
 # along with Openplotter. If not, see <http://www.gnu.org/licenses/>.
 
-import wx, subprocess
-from classes.add_NMEA_0183 import addNMEA_0183
-from classes.paths import Paths
-from classes.op_conf import Conf
-from classes.language import Language
+import wx, os, sys, subprocess
+
+op_folder = os.path.join(os.path.dirname(os.path.realpath(__file__)), '..')
+sys.path.append(op_folder+'/classes')
+from conf import Conf
+from language import Language
+from add_NMEA_0183 import addNMEA_0183
 
 class MyFrame(wx.Frame):
 		
 		def __init__(self):
 
-			self.paths=Paths()
+			self.conf = Conf()
+			self.home = self.conf.home
+			self.currentpath = self.home+self.conf.get('GENERAL', 'op_folder')+'/openplotter'
 
-			self.conf=Conf()
-
-			Language(self.conf.get('GENERAL','lang'))
+			Language(self.conf)
 
 			wx.Frame.__init__(self, None, title=_('NMEA 0183 generator'), size=(690,350))
 			
 			self.SetFont(wx.Font(10, wx.FONTFAMILY_DEFAULT, wx.FONTSTYLE_NORMAL, wx.FONTWEIGHT_NORMAL))
 			
-			self.icon = wx.Icon(self.paths.op_path+'/openplotter.ico', wx.BITMAP_TYPE_ICO)
+			self.icon = wx.Icon(self.currentpath+'/openplotter.ico', wx.BITMAP_TYPE_ICO)
 			self.SetIcon(self.icon)
 
 			wx.StaticBox(self, label=_(' NMEA 0183 '), size=(670, 230), pos=(10, 10))
@@ -51,6 +53,9 @@ class MyFrame(wx.Frame):
 
 			self.delete_nmea_button =wx.Button(self, label=_('delete'), pos=(585, 65))
 			self.Bind(wx.EVT_BUTTON, self.delete_nmea, self.delete_nmea_button)
+
+			self.compat_nmea_button =wx.Button(self, label=_('opencpn\ndefault'), pos=(585, 189))
+			self.Bind(wx.EVT_BUTTON, self.compat_nmea, self.compat_nmea_button)
 
 			self.diagnostic_nmea_button=wx.Button(self, label=_('NMEA Diagnostic'), pos=(10, 250))
 			self.Bind(wx.EVT_BUTTON, self.kplex_diagnostic, self.diagnostic_nmea_button)
@@ -70,7 +75,7 @@ class MyFrame(wx.Frame):
 
 		def start_d(self):
 			subprocess.call(['pkill', '-f', 'SK-base_d.py'])
-			subprocess.Popen(['python',self.paths.op_path+'/SK-base_d.py'])
+			subprocess.Popen(['python',self.currentpath+'/SK-base_d.py'])
 			
 
 		def read_sentences(self):
@@ -99,7 +104,8 @@ class MyFrame(wx.Frame):
 			self.edit_add_nmea(0)
 
 		def edit_add_nmea(self,edit):
-			dlg = addNMEA_0183(edit)
+			self.SetStatusText('')
+			dlg = addNMEA_0183(edit, self.conf)
 			res = dlg.ShowModal()
 			if res == wx.ID_OK:
 				nmea=dlg.nmea
@@ -130,6 +136,7 @@ class MyFrame(wx.Frame):
 			dlg.Destroy()
 
 		def delete_nmea(self,e):
+			self.SetStatusText('')
 			selected_sentence=self.list_nmea.GetFirstSelected()
 			if selected_sentence==-1: 
 				self.SetStatusText('Select a sentence to delete.')
@@ -139,13 +146,20 @@ class MyFrame(wx.Frame):
 			self.conf.set('NMEA0183', 'sentences', str(self.sentences))
 			self.start_d()
 
-
+		def compat_nmea(self,e):
+			if self.list_nmea.GetItemCount()>0:
+				self.SetStatusText(_('This function is only allowed when the list is empty.'))
+				return
+			self.conf.set('NMEA0183', 'sentences', "[['HDG', [['navigation.headingMagnetic', 'x.x|deg', '+', 0.0], '', '', '', ''], 0.5], ['XDR', ['A', ['navigation.attitude.roll', 'x.x|deg', '+', 0.0], 'D', 'ROLL'], 1.0], ['XDR', ['A', ['navigation.attitude.pitch', 'x.x|deg', '+', 0.0], 'D', 'PTCH'], 1.0], ['XDR', ['P', ['environment.outside.pressure', 'x.xxxx', '/', 100000.0], 'B', 'Barometer'], 5.0], ['XDR', ['C', ['environment.outside.temperature', 'x.x|C', '+', 0.0], 'C', 'ENV_OUTAIR_T'], 5.0], ['MTW', [['environment.water.temperature', 'x.x|C', '+', 0.0], 'C'], 5.0]]")
+			self.start_d()
+			self.read_sentences()
+			
 		def kplex_diagnostic(self,e):
 			wx.MessageBox('use diagnostic on NMEA0183\nselect system\npush diagnostic', 'Info', wx.OK | wx.ICON_INFORMATION)
 
 		def sk_diagnostic(self,e):
 			subprocess.call(['pkill', '-f', 'diagnostic-SK-input.py'])
-			subprocess.Popen(['python',self.paths.op_path+'/diagnostic-SK-input.py'])
+			subprocess.Popen(['python',self.currentpath+'/diagnostic-SK-input.py'])
 
 app = wx.App()
 MyFrame().Show()