+++ tools/analog_ads1115.py
@@ -15,23 +15,22 @@
 # You should have received a copy of the GNU General Public License
 # along with Openplotter. If not, see <http://www.gnu.org/licenses/>.
 
-import socket, time, RTIMU, math, csv, datetime, subprocess, sys
-from classes.paths import Paths
-from classes.ads1115 import Ads1115
-from classes.conf_analog import Conf_analog
+import socket, time, datetime, subprocess, sys, os
 
-paths=Paths()
-toolspath=paths.op_path + '/tools'
-home=paths.home
+op_folder = os.path.join(os.path.dirname(os.path.realpath(__file__)), '..')
+sys.path.append(op_folder+'/classes')
+from ads1115 import Ads1115
+from conf_analog import Conf_analog
+
+conf_analog=Conf_analog()
+home = conf_analog.home
 
 if len(sys.argv)>1:
 	if sys.argv[1]=='settings':
-		print toolspath+'openplotter_analog.conf'
-		subprocess.Popen(['leafpad',toolspath+'/openplotter_analog.conf'])
+		print home+'/.openplotter/openplotter_analog.conf'
+		subprocess.Popen(['leafpad',home+'/.openplotter/openplotter_analog.conf'])
 	exit
 else:
-
-	conf_analog=Conf_analog()
 	ads1115=Ads1115()
 
 	a_index = [0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15]