+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+v0.15.0
+* Fix compass performance
+* add ujson package
+* update tools README
+* Fix issue with actions and I2C
+* Fix issue with signal k keys in MQTT
+
+v0.14.5
+* Fix updates from 0.10.0
+* Fix start/stop all actions
+* Add OpenCPN default sentences to NMEA 0183 generator
+* Put config files into ~/.openplotter
+* Put custom tools into ~/.openplotter/tools
+* Create a demo custom tool
+* Remove upgrade raspbianfrom menu update
+* Fix some language issues
+* Optimize conf file management
+* Optimize paths and make them relative to an option in conf file
+* Force update if OP is upto date
+
+v0.14.4
+* Make updates standalone
+
+v0.14.3
+* Fix error afdding I2C sensors
+* Fix SignalK percent to ratio (0-100 -> 0-1) 
+
+v0.14.2
+* Fix error when IMU is not defined
+
+v0.14.1
+* Remove deviation table from Compass tab
+
+v0.14.0
+* New IMU management, add Compass tab
+
 v0.13.1
 * Remove real VNC server and reinstall xrdp